davinci-code/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=13.0
langgraph>=0.0.50
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
google-genai>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
PyYAML>=6.0
rich>=13.7.0
elevenlabs>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.9
groq>=0.9.0
